refactor: 뉴스레터 읽기 페이지를 _bombom 라우트에서 외부로 분리
- pc 버전을 제외하고는 헤더가 다르게 적용되어야 하는 점을 고려하여, 헤더 내부의 분기 처리보다는 페이지를 외부로 분리해 header를 페이지에서 선택할 수 있게 개선함
