+++ frontend/src/pages/detail/components/DetailPageHeader/DetailPageHeader.tsx
